Workaround for watches don't get dropped.
When an HTTP client starts a watch for a proc, and then subsequently disconnects,
the Stream/Receiver channel doesn't get dropped until the proc sends a ProcEvent,
at which point actix tries to write it to the socket, and realizes it's closed.
To work around this, we send a ping event every 2 seconds to every watch.